+++ .gitignore
@@ -3,19 +3,11 @@ node_modules
 jspm_packages
 npm-debug.log
 debug.log
-<<<<<<< HEAD
--app/**/*.js
-=======
 /app/*.js
 /app/**/*.js
->>>>>>> master
 !systemjs.config.extras.js
 !systemjs.config.js
 *.js.map
 _test-output
 _temp
-.DS_Store<<<<<<< HEAD
 .DS_Store
-=======
-.DS_Store
->>>>>>> master
