+++ app/components/dropdown.component.ts
@@ -1,7 +1,7 @@
 ﻿import { Component, Input, Output, EventEmitter, SimpleChanges, forwardRef } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR, FormControl } from '@angular/forms';
 
-
+import { BaseControlValueAccessor } from './base-control-value-accessor'
 
 @Component({
     selector: 'drop-down',
@@ -15,22 +15,17 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR, FormControl } from '@angular/f
     ]
 })
 
-export class DropDownComponent implements ControlValueAccessor {
+export class DropDownComponent extends BaseControlValueAccessor {
 
     @Input() public types: any[] = new Array<any>();
     @Input() public selectedItem: any;
     @Input() public label: string = "";
     @Input() public pkField: string = "";
     @Input() public textField: string = "";
-    @Input() public isReadOnly: boolean = false;
-
-    @Output() public onChange: EventEmitter<any> = new EventEmitter<any>();
-
-
     
 
     constructor() {
-
+        super();
     }
 
     ngOnInit() {
@@ -55,23 +50,6 @@ export class DropDownComponent implements ControlValueAccessor {
         };
     }
 
-    //**************  ControlValueAccessor Methods ****************
-
-    public writeValue(value: any) {
-        console.log(value);
-    }
-
-    public propagateChange = (_: any) => { };
-
-    public registerOnChange(fn: any) {
-        this.propagateChange = fn;
-    }
-
-    public registerOnTouched(fn: any) { }
-
-    public setDisabledState(isDisabled?: boolean): void {
-
-    }
 
     //*************************************************************
