+++ app/components/validation-message.html
@@ -1,4 +1,4 @@
-﻿<div *ngIf="control && control.invalid && (control.dirty || control.touched)" class="alert alert-danger">
+﻿<div *ngIf="isInvalid" class="alert alert-danger">
     <div *ngIf="control.errors.required">
         {{ errorMessage }}
     </div>
