+++ app/files/campaign.json
@@ -1,6 +1,6 @@
 ﻿[
   {
-    "SubCampaignCodeTypeKey": 1,
+    "CampaignCodeTypeKey": 1,
     "ShortName": "90070",
     "LongName": "90070",
     "ShortDesc": "",
@@ -10,7 +10,7 @@
     "SortOrder": 1
   },
   {
-    "SubCampaignCodeTypeKey": 2,
+    "CampaignCodeTypeKey": 2,
     "ShortName": "98941",
     "LongName": "98941",
     "ShortDesc": "",
@@ -20,7 +20,7 @@
     "SortOrder": 2
   },
   {
-    "SubCampaignCodeTypeKey": 3,
+    "CampaignCodeTypeKey": 3,
     "ShortName": "88556",
     "LongName": "88556",
     "ShortDesc": "",
@@ -30,7 +30,7 @@
     "SortOrder": 3
   },
   {
-    "SubCampaignCodeTypeKey": 4,
+    "CampaignCodeTypeKey": 4,
     "ShortName": "78910",
     "LongName": "10987",
     "ShortDesc": "test",
