--- app/services/data.service.js
@@ -1,51 +1,0 @@
-"use strict";
-var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
-    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
-    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
-    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
-    return c > 3 && r && Object.defineProperty(target, key, r), r;
-};
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
-var core_1 = require("@angular/core");
-var http_1 = require("@angular/http");
-require("rxjs/add/observable/of");
-require("rxjs/add/operator/map");
-require("rxjs/add/operator/catch");
-var ApiDataService = (function () {
-    function ApiDataService(http) {
-        this.http = http;
-    }
-    ApiDataService.prototype.getStatesTypes = function () {
-        return this.http.get("app/files/states.json")
-            .map(function (res) { return res.json(); })
-            .catch(function (error, caught) {
-            console.log(error);
-            return error;
-        });
-    };
-    ApiDataService.prototype.getSubCampaignCodeTypes = function () {
-        return this.http.get("app/files/campaign.json")
-            .map(function (res) { return res.json(); })
-            .catch(function (error, caught) {
-            console.log(error);
-            return error;
-        });
-    };
-    ApiDataService.prototype.getAdminTypes = function () {
-        return this.http.get("app/files/admins.json")
-            .map(function (res) { return res.json(); })
-            .catch(function (error, caught) {
-            console.log(error);
-            return error;
-        });
-    };
-    return ApiDataService;
-}());
-ApiDataService = __decorate([
-    core_1.Injectable(),
-    __metadata("design:paramtypes", [http_1.Http])
-], ApiDataService);
-exports.ApiDataService = ApiDataService;
-//# sourceMappingURL=data.service.js.map
